Getting it to work with ActiveModel 3.2 closes #131
* Look like the specs runs faster with ActiveModel 3.2.

* There is small change on how the dirty activemodel feature works.
   The previous value is now only available after the model hasn been saved.
   This is probably the reason why one versioning spec did not work (set to pending).

+++ lib/neo4j/rails/persistence.rb
@@ -237,9 +237,9 @@ module Neo4j
         allow_destroy, reject_if = [options[:allow_destroy], options[:reject_if]] if options
         begin
           # Check if we want to destroy not found nodes (e.g. {..., :_destroy => '1' } ?
-          destroy = allow_destroy && attr[:_destroy] && attr[:_destroy] != '0'
+          destroy = attr.delete(:_destroy)
           found = _find_node(rel_type, attr[:id]) || Model.find(attr[:id])
-          if destroy
+          if allow_destroy && destroy && destroy != '0'
             found.destroy if found
           else
             if not found
